+++ src/main/java/com/github/kaaz/emily/economy/MoneyTreeCommand.java
@@ -17,7 +17,7 @@ public class MoneyTreeCommand extends AbstractCommand {
         super("moneytree", BotRole.BOT_ADMIN, ModuleLevel.ECONOMY, "mt", null, "Gives free money");
     }
     @Command
-    public void handle(@Argument Float amount, @Argument Configurable configurable){
+    public void handle(@Argument Integer amount, @Argument Configurable configurable){
         ConfigHandler.changeSetting(CurrentMoneyConfig.class, configurable, aFloat -> aFloat + amount);
     }
 }
