+++ src/main/java/com/github/kaaz/emily/economy/configs/CurrentMoneyConfig.java
@@ -7,8 +7,8 @@ import com.github.kaaz.emily.perms.BotRole;
 /**
  * Made by nija123098 on 5/1/2017.
  */
-public class CurrentMoneyConfig extends AbstractConfig<Float, Configurable> {
+public class CurrentMoneyConfig extends AbstractConfig<Integer, Configurable> {
     public CurrentMoneyConfig() {
-        super("current_money", BotRole.BOT_ADMIN, 0F, "The amount of money a guild user has");
+        super("current_money", BotRole.BOT_ADMIN, 0, "The amount of money a guild user has");
     }
 }
