+++ src/main/java/com/github/kaaz/emily/economy/configs/StartingMoneyConfig.java
@@ -12,14 +12,13 @@ import com.github.kaaz.emily.perms.BotRole;
 /**
  * Made by nija123098 on 5/1/2017.
  */
-public class StartingMoneyConfig extends AbstractConfig<Float, Configurable> {
+public class StartingMoneyConfig extends AbstractConfig<Integer, Configurable> {
     public StartingMoneyConfig() {
-        super("starting_money", BotRole.BOT_ADMIN, 0F, "The amount of money a config starts with");
+        super("starting_money", BotRole.BOT_ADMIN, 0, "The amount of money a config starts with");
     }
     @EventListener
     public void handle(DiscordUserJoin event){
         GuildUser guildUser = GuildUser.getGuildUser(event.getGuild(), event.getUser());
-        if (Math.abs(event.getJoinTime() - GuildUserJoinTimeConfig.get(guildUser)) < 5000)
-            ConfigHandler.changeSetting(CurrentMoneyConfig.class, guildUser, integer -> this.getValue(event.getGuild()));
+        if (Math.abs(event.getJoinTime() - GuildUserJoinTimeConfig.get(guildUser)) < 5000) ConfigHandler.changeSetting(CurrentMoneyConfig.class, guildUser, integer -> this.getValue(event.getGuild()));
     }// integer should be 0, but the change is for safety
 }
